refactor: split riscv codegen
