+++ src/backend/riscv_codegen.cpp
@@ -4,218 +4,7 @@
 #include "util.h"
 #include <cassert>
 #include <cstddef>
-
-int AddrManager::getNextId() { return id_counter++; }
-
-std::string AddrManager::idToAddr(int id) {
-  if (id_addr_map.find(id) == id_addr_map.end()) {
-    id_addr_map[id] = getReg();
-  }
-  return id_addr_map[id];
-}
-
-std::string AddrManager::getReg() {
-  auto reg = regs.begin();
-  std::string reg_str = *reg;
-  regs.erase(reg);
-  return reg_str;
-}
-
-template <typename T>
-std::string AddrManager::getAddrImpl(const T &obj,
-                                     std::unordered_map<const T *, int> &map) {
-  int id;
-  if (map.find(&obj) == map.end()) {
-    id = getNextId();
-    map[&obj] = id;
-  } else {
-    id = map[&obj];
-  }
-  return idToAddr(id);
-}
-
-void AddrManager::freeReg(const std::string &reg) {
-  if (reg == "x0") {
-    return;
-  }
-  if (reg[0] == 'a') {
-    return;
-  }
-  regs.insert(reg);
-}
-
-std::string AddrManager::getAddr(const koopa_raw_binary_t &binary) {
-  return getAddrImpl(binary, raw_bin_id_map);
-}
-
-std::string AddrManager::getAddr(const koopa_raw_store_t &store) {
-  return getAddrImpl(store, raw_store_id_map);
-}
-
-std::string AddrManager::getAddr(const koopa_raw_load_t &load) {
-  return getAddrImpl(load, raw_load_id_map);
-}
-
-std::string AddrManager::getAddr(const koopa_raw_value_t &value) {
-  if (value->kind.tag == KOOPA_RVT_INTEGER &&
-      value->kind.data.integer.value == 0) {
-    return "x0";
-  }
-  if (value->kind.tag == KOOPA_RVT_BINARY) {
-    return getAddr(value->kind.data.binary);
-  }
-
-  return getAddrImpl(value, raw_val_id_map);
-}
-
-void AddrManager::freeId(const koopa_raw_value_t &value) {
-  raw_val_id_map.erase(&value);
-}
-
-void AddrManager::freeId(const koopa_raw_store_t &store) {
-  raw_store_id_map.erase(&store);
-}
-
-void AddrManager::freeId(const koopa_raw_load_t &load) {
-  raw_load_id_map.erase(&load);
-}
-
-void AddrManager::freeId(const koopa_raw_binary_t &binary) {
-  raw_bin_id_map.erase(&binary);
-}
-
-int StackOffsetManager::getNextId() { return ++id_counter; }
-
-int StackOffsetManager::id_to_offset(int id) {
-  if (id_to_offset_map.find(id) == id_to_offset_map.end()) {
-    assert(false);
-  }
-  return id_to_offset_map[id] + a;
-}
-
-void StackOffsetManager::setOffset(const koopa_raw_value_t &value) {
-  auto kind = value->kind.tag;
-  switch (kind) {
-  case KOOPA_RVT_LOAD:
-    load_id_map[&value->kind.data.load] = getNextId();
-    id_to_offset_map[load_id_map[&value->kind.data.load]] =
-        current_stack_offset;
-    current_stack_offset += 4;
-    break;
-  case KOOPA_RVT_BINARY:
-    binary_id_map[&value->kind.data.binary] = getNextId();
-    id_to_offset_map[binary_id_map[&value->kind.data.binary]] =
-        current_stack_offset;
-    current_stack_offset += 4;
-    break;
-  case KOOPA_RVT_ALLOC:
-    alloc_name_id_map[value->name] = getNextId();
-    id_to_offset_map[alloc_name_id_map[value->name]] = current_stack_offset;
-    current_stack_offset += 4;
-    break;
-  case KOOPA_RVT_FUNC_ARG_REF:
-    func_arg_idx_id_map[value->kind.data.func_arg_ref.index] = getNextId();
-    id_to_offset_map[func_arg_idx_id_map[value->kind.data.func_arg_ref.index]] =
-        current_stack_offset;
-    current_stack_offset += 4;
-    break;
-  case KOOPA_RVT_CALL:
-    call_id_map[&value->kind.data.call] = getNextId();
-    id_to_offset_map[call_id_map[&value->kind.data.call]] =
-        current_stack_offset;
-    current_stack_offset += 4;
-    break;
-  default:
-    assert(false);
-  }
-}
-
-void StackOffsetManager::setOffset(int idx, int offset) {
-  func_arg_idx_id_map[idx] = getNextId();
-  id_to_offset_map[func_arg_idx_id_map[idx]] = offset;
-}
-
-int StackOffsetManager::getOffset(const koopa_raw_load_t &load) {
-  int id = 0;
-  if (load_id_map.find(&load) == load_id_map.end()) {
-    assert(false);
-  } else {
-    id = load_id_map[&load];
-  }
-  return id_to_offset(id);
-}
-
-int StackOffsetManager::getOffset(const koopa_raw_binary_t &binary) {
-  int id = 0;
-  if (binary_id_map.find(&binary) == binary_id_map.end()) {
-    assert(false);
-  } else {
-    id = binary_id_map[&binary];
-  }
-  return id_to_offset(id);
-}
-
-int StackOffsetManager::getOffset(const std::string &alloc_name) {
-  int id = 0;
-  if (alloc_name_id_map.find(alloc_name) == alloc_name_id_map.end()) {
-    assert(false);
-  } else {
-    id = alloc_name_id_map[alloc_name];
-  }
-  return id_to_offset(id);
-}
-
-int StackOffsetManager::getOffset(
-    const koopa_raw_func_arg_ref_t &func_arg_ref) {
-  int id = 0;
-  if (func_arg_idx_id_map.find(func_arg_ref.index) ==
-      func_arg_idx_id_map.end()) {
-    assert(false);
-  } else {
-    id = func_arg_idx_id_map[func_arg_ref.index];
-  }
-  return id_to_offset(id);
-}
-
-int StackOffsetManager::getOffset(const koopa_raw_call_t &call) {
-  int id = 0;
-  if (call_id_map.find(&call) == call_id_map.end()) {
-    assert(false);
-  } else {
-    id = call_id_map[&call];
-  }
-  return id_to_offset(id);
-}
-
-int StackOffsetManager::getOffset(const koopa_raw_value_t &value) {
-  switch (value->kind.tag) {
-  case KOOPA_RVT_BINARY:
-    return getOffset(value->kind.data.binary);
-    break;
-  case KOOPA_RVT_FUNC_ARG_REF:
-    return getOffset(value->kind.data.func_arg_ref);
-    break;
-  case KOOPA_RVT_LOAD:
-    return getOffset(value->kind.data.load);
-    break;
-  case KOOPA_RVT_CALL:
-    return getOffset(value->kind.data.call);
-    break;
-  default:
-    assert(false);
-  }
-  assert(false);
-  return 0;
-}
-
-void StackOffsetManager::clear() {
-  load_id_map.clear();
-  binary_id_map.clear();
-  alloc_name_id_map.clear();
-  id_to_offset_map.clear();
-  current_stack_offset = 0;
-  id_counter = 0;
-}
+#include <algorithm>
 
 CodeGen::CodeGen(const std::string &koopa_ir) {
   push_addr_manager();
